fix: MPS disabled due to PyG 2.6.1 GINConv bug, keep parallel CPU approach
PyG's GINConv.propagate() creates x_r on CPU even when model/data on MPS.
Known issue in PyG <2.7. Workaround: CPU + 4 parallel processes.
Also: strip_subgraph utility, .float() cast for ZINC integer features.

+++ src/models/base_gnn.py
@@ -30,7 +30,7 @@ class GINEncoder(nn.Module):
         if single:
             data = Batch.from_data_list([data])
 
-        x, edge_index, batch = data.x, data.edge_index, data.batch
+        x, edge_index, batch = data.x.float(), data.edge_index, data.batch
 
         for conv, bn in zip(self.convs, self.bns):
             x = conv(x, edge_index)
